requirements.txt
crewai>=0.108.0
gradio>=5.0
litellm>=1.50
pyyaml
